barbearia-backend/requirements.txt
# requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

pytest==7.4.3
PyJWT==2.8.0
firebase-admin==6.5.0
python-dotenv==1.0.0